update tracking share

+++ apps/metvuong/frontend/web/themes/mv_desktop1/views/ad/_partials/shareSocial.php
@@ -22,12 +22,14 @@ use yii\helpers\Url;
                             <ul class="clearfix">
                                 <li>
                                     <a href="#"
-                                       data-url="<?=Url::to(['/ad/tracking-share-facebook', 'product_id' => $product_id], true)?>" class="share-facebook">
+                                       data-url="<?=Url::to(['/ad/tracking-share', 'product_id' => $product_id, 'type' => \vsoft\tracking\models\base\AdProductShare::SHARE_FACEBOOK], true)?>" class="share-facebook">
                                         <div class="circle"><div><span class="icon icon-face"></span></div></div>
                                     </a>
                                 </li>
                                 <li>
-                                    <a href="#" data-toggle="modal" data-target="#popup-email" class="email-btn">
+                                    <a href="#"
+                                       data-url="<?=Url::to(['/ad/tracking-share', 'product_id' => $product_id, 'type' => \vsoft\tracking\models\base\AdProductShare::SHARE_EMAIL], true)?>"
+                                       data-toggle="modal" data-target="#popup-email" class="email-btn">
                                         <div class="circle"><div><span class="icon icon-email-1"></span></div></div>
                                     </a>
                                 </li>
@@ -47,8 +49,8 @@ use yii\helpers\Url;
         window.open('http://www.facebook.com/sharer.php?s=100&p[url]=' + url + '&p[title]=' + title + '&p[summary]=' + descr + '&p[images][0]=' + image, 'sharer', 'top=' + winTop + ',left=' + winLeft + ',toolbar=0,status=0,width=' + winWidth + ',height=' + winHeight);
     }
 
-    $('.share-facebook').click(function (){
-        var url = $('.share-facebook').data("url");
+    $('#popup-share-social ul li a').click(function (){
+        var url = $(this).data("url");
         if(url.length > 0) {
             $('body').loading();
             $.ajax({
@@ -56,10 +58,12 @@ use yii\helpers\Url;
                 dataType: 'json',
                 url: url,
                 success: function (data) {
+                    $('#popup-share-social').modal('hide');
                 }
             });
             $('body').loading({done:true});
-            fbShare('<?=$url ?>', '<?=$title ?>', '<?=StringHelper::truncate($description, 120) ?>', '<?= $image ?>', 520, 350);
+            if($(this).attr('class') == 'share-facebook')
+                fbShare('<?=$url ?>', '<?=$title ?>', '<?=StringHelper::truncate($description, 120) ?>', '<?= $image ?>', 520, 350);
 
             return true;
         }
